python-flask/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
Flask-Limiter>=3.5.0,<4.0.0
pymongo>=4.6.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
PyJWT>=2.8.0,<3.0.0
passlib>=1.7.4,<2.0.0
email-validator>=2.1.0,<3.0.0
